+++ apps/www/app/(private)/inbox/components/received/received-card.tsx
@@ -3,11 +3,7 @@ import { cn } from "@umamin/ui/lib/utils";
 import { formatDistanceToNow } from "date-fns";
 import { ReceivedMessageMenu } from "./received-card-menu";
 
-export function ReceivedMessageCard({
-  data,
-}: {
-  data: SelectMessage & { canBlock?: boolean };
-}) {
+export function ReceivedMessageCard({ data }: { data: SelectMessage }) {
   return (
     <div id={`umamin-${data.id}`}>
       <div
@@ -18,7 +14,6 @@ export function ReceivedMessageCard({
         <div className="absolute top-4 right-4 text-muted-foreground">
           <ReceivedMessageMenu
             {...data}
-            canBlock={data.canBlock}
             reply={data.reply}
             updatedAt={data.updatedAt}
           />