+++ apps/www/lib/query.ts
@@ -27,6 +27,7 @@ export const queryKeys = {
     ["user-profile-viewer", username] as const,
   followers: (username: string) => ["followers", username] as const,
   following: (username: string) => ["following", username] as const,
+  blockedUsers: () => ["blocked_users"] as const,
   receivedMessages: () => ["received_messages"] as const,
   sentMessages: () => ["sent_messages"] as const,
   notifications: () => ["notifications"] as const,