+++ packages/db/migrations/0009_heavy_mephistopheles.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `user` ADD `blocked_words` text;